+++ lib/src/commands/update_command.dart
@@ -55,7 +55,7 @@ class UpdateCommand extends Command<int> {
       _logger
         ..info('')
         ..info(
-          'See the release notes here: https://github.com/iamngoni/gitwhisper/releases/tag/$latestVersion',
+          'See the change notes here: https://pub.dev/packages/gitwhisper/changelog#$latestVersion',
         );
 
       return ExitCode.success.code;